lohas-papers-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
redis==5.2.1
sqlalchemy[asyncio]==2.0.37
asyncpg==0.30.0
anthropic==0.44.0
python-dotenv==1.0.1
